Project Development Phase/Sprint 4/requirements.txt
APScheduler==3.9.1.post1
backports.zoneinfo
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-APScheduler==1.12.4
ibm-db
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.2.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0